requirements.txt
APScheduler==3.6.3
backports.entry-points-selectable==1.1.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.6
distlib==0.3.3
filelock==3.3.0
greenlet==1.1.2
idna==3.2
kaleido==0.2.1
pbr==5.6.0
platformdirs==2.4.0
plotly==5.3.1
python-telegram-bot==13.7
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
stevedore==3.4.0
tenacity==8.0.1
tzlocal==3.0
urllib3==1.26.7
virtualenv==20.8.1
virtualenv-clone==0.5.7
schedule~=1.1.0
prettytable~=2.4.0